Move acme tests to binary test package
